+++ src/LoginRadiusSDK/CustomerRegistration/Authentication/PINAuthenticationAPI.php
@@ -23,7 +23,7 @@ class PINAuthenticationAPI extends Functions
 
 
     /**
-     * This API is used to login a user by pin and session_token.
+     * This API is used to login a user by pin and session token.
      * @param loginByPINModel Model Class containing Definition of payload for LoginByPin API
      * @param sessionToken Session Token of user
      * @return Response containing User Profile Data and access token