+++ src/LoginRadiusSDK/CustomerRegistration/Authentication/PhoneAuthenticationAPI.php
@@ -259,7 +259,7 @@ class PhoneAuthenticationAPI extends Functions
 
 
     /**
-     * This API is used to delete the Phone ID on a user's account via the access_token
+     * This API is used to delete the Phone ID on a user's account via the access token
      * @param accessToken Uniquely generated identifier key by LoginRadius that is activated after successful authentication.
      * @return Response containing Definition of Delete Request
      * 11.7